Set language preference in baseMapActivity and baseRevealregisterActivity
