+++ opensrp-reveal/src/main/java/org/smartregister/reveal/view/BaseMapActivity.java
@@ -1,7 +1,11 @@
 package org.smartregister.reveal.view;
 
+import android.content.Context;
 import android.os.Bundle;
 
+import org.smartregister.reveal.BuildConfig;
+import org.smartregister.reveal.util.Country;
+import org.smartregister.util.LangUtils;
 import org.smartregister.view.activity.MultiLanguageActivity;
 
 /**
@@ -59,5 +63,15 @@ public abstract class BaseMapActivity extends MultiLanguageActivity {
         super.onDestroy();
         if (kujakuMapView != null)
             kujakuMapView.onDestroy();
+    }
+
+    @Override
+    protected void attachBaseContext(Context base) {
+        if (BuildConfig.BUILD_COUNTRY == Country.THAILAND) {
+            LangUtils.saveLanguage(base.getApplicationContext(), "th");
+        } else {
+            LangUtils.saveLanguage(base.getApplicationContext(), "en");
+        }
+        super.attachBaseContext(base);
     }
 }
